% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dotermgenes}
\alias{dotermgenes}
\title{All DO term annotated genes.}
\format{
An character array with 15106 elements:
}
\usage{
dotermgenes
}
\description{
A dataset includes 15106 genes.
}
\keyword{datasets}
