% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_package.R
\docType{package}
\name{EnrichDO-package}
\alias{EnrichDO-package}
\alias{EnrichDO}
\title{EnrichDO
Enrichment analyses including a variety of statistical models and visualization schemes for discovering the disease-gene relationship under biological big data.}
\description{
To implement disease ontology (DO) enrichment analysis, this package is designed and presents a double weighted model based on the latest annotations of the human genome with DO terms, by integrating the DO graph topology on a global scale. This package exhibits high accuracy that it can identify more specific DO terms, which alleviates the over enriched problem. The package includes various statistical models and visualization schemes for discovering the associations between genes and diseases from biological big data.
}
\author{
Liang cheng, Haixiu Yang, Hongyu Fu

Maintainer: Haixiu Yang \email{yanghaixiu@ems.hrbmu.edu.cn}
}
\keyword{package}
