% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_class.R
\docType{class}
\name{EnrichResult-class}
\alias{EnrichResult-class}
\title{Class 'EnrichResult'
This class represents the result of enrich analysis}
\description{
Class 'EnrichResult'
This class represents the result of enrich analysis
}
\section{Slots}{

\describe{
\item{\code{enrich}}{a data frame of enrichment result}

\item{\code{test}}{Statistical test}

\item{\code{method}}{Multiple test correction methods}

\item{\code{m}}{the maximum number of ancestor layers for ontology enrichment}

\item{\code{maxGsize}}{The maximum number of DOTerm genes in enrichment analysis}

\item{\code{minGsize}}{The minimum number of DOTerm genes in enrichment analysis}

\item{\code{traditional}}{Indicates whether the traditional ORA method is used}

\item{\code{delta}}{The highest p-value of significance for each node}

\item{\code{penalize}}{Whether to use penalty function in enrichment analysis}

\item{\code{interestGenes}}{A valid interest gene set}
}}

\author{
Haixiu Yang
}
\keyword{classes}
