% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_method.R
\docType{methods}
\name{EnrichTab}
\alias{EnrichTab}
\alias{EnrichTab,EnrichResult-method}
\title{Extraction of enrichment results}
\usage{
EnrichTab(object, all)

\S4method{EnrichTab}{EnrichResult}(object, all = TRUE)
}
\arguments{
\item{object}{EnrichResult}

\item{all}{Logical values, TRUE to extract all enrich results and FALSE to extract only significant results.}
}
\value{
data.frame
}
\description{
EnrichTab method for \code{EnrichResult} instance
}
\author{
Haixiu Yang
}
