% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_method.R
\name{show,EnrichResult-method}
\alias{show,EnrichResult-method}
\title{show method}
\usage{
\S4method{show}{EnrichResult}(object)
}
\arguments{
\item{object}{A \code{EnrichResult} instance.}
}
\value{
print info
}
\description{
show method for \code{EnrichResult} instance
}
\examples{
demo.data <- c(1636, 351, 102)
res <- doEnrich(interestGenes = demo.data, traditional = TRUE)
show(res)
}
\author{
Haixiu Yang
}
