% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{showDoTerms}
\alias{showDoTerms}
\title{showDoTerms}
\usage{
showDoTerms(doterms = doterms)
}
\arguments{
\item{doterms}{a data frame of DOterms.}
}
\value{
text
}
\description{
show DOterms
}
\examples{
mock_doterms <- data.frame(
  DOID = c("DOID:0000001", "DOID:0000002"),
  DOTerm = c("disease 1", "disease 2")
)
showDoTerms(mock_doterms)
}
\author{
Haixiu Yang
}
