% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_method.R
\name{viewDetailResult}
\alias{viewDetailResult}
\alias{viewDetailResult,EnrichResult-method}
\title{viewDetailResult method}
\usage{
viewDetailResult(object, ...)

\S4method{viewDetailResult}{EnrichResult}(object)
}
\arguments{
\item{object}{A \code{EnrichResult} instance.}

\item{...}{Additional arguments.}
}
\value{
data.frame of complete enrichment results
}
\description{
viewDetailResult method for \code{EnrichResult} instance
}
\examples{
# Assuming 'weighted_demo_result' is an EnrichResult object generated by doEnrich()
demo.data <- c(1636, 351, 102, 2932, 3077)
weighted_demo_result <- doEnrich(interestGenes=demo.data,
                        test="fisherTest",
                        method="holm",
                        m=1,
                        minGsize=10,
                        maxGsize=2000,
                        delta=0.05,
                        penalize=TRUE)
detail_df <- viewDetailResult(weighted_demo_result)
head(detail_df)
}
\author{
Haixiu Yang
}
