% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_method.R
\name{viewSummaryResult}
\alias{viewSummaryResult}
\alias{viewSummaryResult,EnrichResult-method}
\title{viewSummaryResult method}
\usage{
viewSummaryResult(object, P = 1, Q = 1, ...)

\S4method{viewSummaryResult}{EnrichResult}(object, P = 1, Q = 1)
}
\arguments{
\item{object}{A \code{EnrichResult} instance.}

\item{P}{pvalue cutoff.}

\item{Q}{p.adjust cutoff.}

\item{...}{Additional arguments.}
}
\value{
data.frame of standardized enrichment results
}
\description{
viewSummaryResult method for \code{EnrichResult} instance
}
\examples{
# Assuming 'weighted_demo_result' is an EnrichResult object generated by doEnrich()
demo.data <- c(1636, 351, 102, 2932, 3077)
weighted_demo_result <- doEnrich(interestGenes=demo.data,
                        test="fisherTest",
                        method="holm",
                        m=1,
                        minGsize=10,
                        maxGsize=2000,
                        delta=0.05,
                        penalize=TRUE)
summary_df <- viewSummaryResult(weighted_demo_result)
head(summary_df)
}
\author{
Haixiu Yang
}
