% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{writeDoTerms}
\alias{writeDoTerms}
\title{writeDoTerms}
\usage{
writeDoTerms(doterms = doterms, file)
}
\arguments{
\item{doterms}{a data frame of do terms.}

\item{file}{the address and name of the output file.}
}
\value{
text
}
\description{
Output DOterms as text
}
\examples{
writeDoTerms(doterms, file=file.path(tempdir(), 'doterms.txt'))
}
\author{
Haixiu Yang
}
