% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{convDraw}
\alias{convDraw}
\title{convDraw}
\usage{
convDraw(resultDO)
}
\arguments{
\item{resultDO}{a data frame of enrichment result}
}
\value{
data.frame
}
\description{
using the result of writeResult for convenience drawing.
}
\examples{
#'#Draw from wrireResult output files
#Firstly, read the wrireResult output file,using the following two lines
data<-read.delim(file.path(system.file("examples", package = "EnrichDO"),"result.txt"))
convDraw(resultDO = data)
#then, Use the drawing function you need
drawGraphViz(enrich=enrich)    #Tree diagram
drawPointGraph(enrich=enrich)  #Bubble diagram
drawBarGraph(enrich=enrich)    #Bar plot
}
\author{
Haixiu Yang
}
