% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEnrich.R
\name{doEnrich}
\alias{doEnrich}
\title{doEnrich}
\usage{
doEnrich(
  interestGenes,
  test = "hypergeomTest",
  method = "BH",
  m = 1,
  maxGsize = 5000,
  minGsize = 5,
  traditional = FALSE,
  delta = 0.01,
  result_do = NULL,
  penalize = T
)
}
\arguments{
\item{interestGenes}{a vector of gene IDs.}

\item{test}{One of "fisherTest","hypergeomTest","binomTest","chisqTest" and "logoddTest" statistical model. Default is hypergeomTest.}

\item{method}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr" and "none",for P value correction.}

\item{m}{Set the maximum number of ancestor layers for ontology enrichment. Default is layer 1.}

\item{maxGsize}{indicates that doterms with more annotation genes than maxGsize are ignored, and the P value of these doterms is set to 1.}

\item{minGsize}{indicates that doterms with less annotation genes than minGsize are ignored, and the P value of these doterms is set to 1.}

\item{traditional}{a logical variable, TRUE for traditional enrichment analysis, FALSE for enrichment analysis with weights. Default is FALSE.}

\item{delta}{Set the threshold of nodes, if the p value of doterm is greater than delta, the nodes are not significant, and these nodes are not weighted.}

\item{result_do}{Receives the file output by the wrireResult function, which is used to visually display the enrichment results (without running the enrichment operation again).}

\item{penalize}{Logical value, whether to add a penalty to the node.Adding a penalty will look for nodes with more branches.}
}
\value{
a data frame of enrichment result
}
\description{
given a list of genes, this function combines topological properties of the disease ontology structure for enrichment analysis.
}
\examples{
#The enrichment results were obtained by using demo.data
demo.data=c(1636,351,102,2932,3077,348,4137,54209,5663,5328,23621,3416,3553)
doEnrich(interestGenes=demo.data)

#setting the penalty to FALSE, the algorithm can mitigate the extent of reduced nodes' weights
doEnrich(interestGenes=demo.data,penalize = FALSE)

#Statistical models and P-value correction can be set
doEnrich(demo.data,test="hypergeomTest",method="holm")

#Using the traditional enrichment analysis method.
doEnrich(demo.data,traditional=TRUE)

#Between the number of genes in minGsize and maxGsize doterm enrichment analysis
doEnrich(demo.data,minGsize=5,maxGsize=500)

#Draw from wrireResult output files
#Firstly, read the wrireResult output file,using the following two lines
#data<-read.delim(yourfile)
#doEnrich(result_do = data)
#then, Use the drawing function you need
drawGraphViz(enrich)    #Tree diagram
drawPointGraph(enrich)  #Bubble diagram
drawBarGraph(enrich)    #Bar plot
}
\author{
Haixiu Yang
}
