% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{drawBarGraph}
\alias{drawBarGraph}
\title{drawBarGraph}
\usage{
drawBarGraph(EnrichResult = NULL, enrich = NULL, n = 10, delta = 1e-15)
}
\arguments{
\item{EnrichResult}{the EnrichResult object}

\item{enrich}{a data frame of enrichment result}

\item{n}{number of bars}

\item{delta}{the threshold of P value}
}
\value{
bar graph
}
\description{
The enrichment results are shown in a bar chart
}
\examples{
demo.data <- c(1636,351,102,2932,3077,348,4137,54209)
sample1 <- doEnrich(interestGenes=demo.data,maxGsize = 100, minGsize=10)
drawBarGraph(EnrichResult=sample1, n=10, delta=0.05)
}
\author{
Haixiu Yang
}
