% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{drawHeatmap}
\alias{drawHeatmap}
\title{drawHeatmap}
\usage{
drawHeatmap(
  interestGenes,
  enrich = enrich,
  DOID_n = 10,
  gene_n = 50,
  fontsize_row = 10,
  readable = TRUE,
  ...
)
}
\arguments{
\item{interestGenes}{A collection of interest genes in vector form}

\item{enrich}{a data frame of the enrichment result}

\item{DOID_n}{There are DOID_n nodes with the highest significance in the enrichment results.}

\item{gene_n}{Among the selected DOID_n nodes, the top gene_n genes with the highest weight sum are selected to show.}

\item{fontsize_row}{Set the font size of the gene tag.}

\item{readable}{Logical value that controls whether the gene tag is in symbol format}

\item{...}{Other parameters in the pheatmap function also apply.}
}
\value{
heat map
}
\description{
The top DOID_n nodes in the enrichment results showed the top gene_n genes with the highest weight sum.
}
\examples{
demo.data=c(1636,351,102,2932,3077,348,4137,54209,5663,5328,23621,3416,3553)
sample6<-doEnrich(interestGenes=demo.data)
enrich<-EnrichTab(object =sample6,all=TRUE)
drawHeatmap(interestGenes=demo.data,enrich = enrich,gene_n = 10)
}
\author{
Haixiu Yang
}
