% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{drawPointGraph}
\alias{drawPointGraph}
\title{drawPointGraph}
\usage{
drawPointGraph(enrich = enrich, n = 10, delta = 1e-15)
}
\arguments{
\item{enrich}{a data frame of enrichment result.}

\item{n}{number of points.}

\item{delta}{the threshold of P value.}
}
\value{
scatter graph
}
\description{
The enrichment results are shown in a scatter plot
}
\examples{
demo.data=c(1636,351,102,2932,3077,348,4137,54209,5663,5328,23621,3416,3553)
sample2<-doEnrich(interestGenes=demo.data)
drawPointGraph(enrich=sample2,n=10,delta=0.05)
}
\author{
Haixiu Yang
}
