% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{showDoTerms}
\alias{showDoTerms}
\title{showDoTerms}
\usage{
showDoTerms(doterms = doterms)
}
\arguments{
\item{doterms}{a data frame of DOterms.}
}
\value{
text
}
\description{
show DOterms
}
\examples{
showDoTerms(doterms)
}
\author{
Haixiu Yang
}
