% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_algorithm.R
\name{TermStruct}
\alias{TermStruct}
\title{Enrich_internal}
\usage{
TermStruct(result_do)
}
\arguments{
\item{result_do}{Receives the file output by the wrireResult function, which is used to visually display the enrichment results (without running the enrichment operation again).}
}
\value{
A \code{enrich} instance.
}
\description{
Internal calculation of enrichment analysis
}
\author{
Haixiu Yang
}
