% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDO_visualization.R
\name{writeResult}
\alias{writeResult}
\title{writeResult}
\usage{
writeResult(enrich = enrich, file, Q = 1, P = 1)
}
\arguments{
\item{enrich}{a data frame of the enrichment result.}

\item{file}{the address and name of the output file.}

\item{Q}{Output only doterm information with p.adjust values less than or equal to Q.}

\item{P}{Output only doterm information with p values less than or equal to P.}
}
\value{
text
}
\description{
Output enrichment result as text
}
\examples{
demo.data=c(1636,351,102,2932,3077,348,4137,54209,5663,5328,23621,3416,3553)
sample4<-doEnrich(interestGenes=demo.data)
writeResult(sample4@enrich,file=file.path(tempdir(),"result.txt"))
}
\author{
Haixiu Yang
}
